Switch to Scott's new CrIS reader, fixed cris_paths.m, removed empty dcc*.m
CrIS: (1) switched to readsdr_rawpd_all.m from readsdr_rawpd.m This
new function reads more attributes, including the name of the geo
file.  Now, the geo file name is obtained from the attributes.  (2)
cris_paths.m in rtp had an error (util related).  (3) p.findex now
forced to be an int32 from the beginning in rtp_core.m (4) cris/utils
now has two routines to handle packing the CrIS QA bits (QA_to_bits.m,
cris_sdr_QAFlags.m).

AIRS: Integrated in Paul's changes that only run "getairs" to get
AIRS QA data when the script is running on a umbc.edu machine.

This version works for AIRS (ecwmf, gfs) on Dec. 18, 2011 and IASI
(ecmwf, gfs) on Dec. 18, 2011 on Strow's leek machine.  It also works
for CrIS (ecmwf) for March 10, 2011.

+++ cris/rtp/rtp_core.m
@@ -69,7 +69,7 @@ for hour = 0:23
     d = dir(f{i});
     disp(['Reading ' f{i}])
     [p pattr]=readsdr_rtp(f{i});
-    p.findex = ones(size(p.rtime)) * i;
+    p.findex = int32(ones(size(p.rtime)) * i);
 
     % Now change indices to g4 of SARTA
     robs = p.robs1;
@@ -89,7 +89,7 @@ for hour = 0:23
     prof(i) = p;
     clear p;
   end % file loop
-
+  
   % if no files were loaded continue on to the next hour
   if ~exist('prof','var')
     continue
@@ -161,11 +161,11 @@ for hour = 0:23
     [head hattr prof pattr] = rtpadd_ecmwf_data(head,hattr,prof,pattr);
   end
 
-
   disp('adding emissivity');
   rtime = rtpget_date(prof,pattr);
   dv = datevec(JOB(1));
   [prof emis_qual emis_str] = Prof_add_emis(prof, dv(1), dv(2), dv(3), 0, 'nearest', 2, 'all');
+ 
 
   [head,hattr,prof,pattr,summary] = rtp_cris_subset(head,hattr,prof,pattr,strcmp(rtpset,'subset'));
   if isempty(prof); disp('ERROR: no data returned'); continue; end  % if no data was returned
